# api/swagger/swagger.yaml
basePath: /
definitions:
  response.Response:
    properties:
      data: {}
      error:
        type: string
      status:
        description: '"success" or "error"'
        type: string
      status_code:
        description: HTTP status code
        type: integer
    type: object
  service.CreateOrderRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/service.OrderItemRequest'
        minItems: 1
        type: array
      note:
        type: string
      order_code:
        type: string
      type:
        enum:
        - IMPORT
        - EXPORT
        type: string
    required:
    - items
    - order_code
    - type
    type: object
  service.CreateUserRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      phone:
        type: string
      role:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - phone
    - role
    - username
    type: object
  service.LoginUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  service.OrderItemRequest:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
      unit_price:
        type: number
    required:
    - product_id
    - quantity
    - unit_price
    type: object
  service.ProductResponse:
    properties:
      current_stock:
        type: integer
      id:
        type: string
      name:
        type: string
      price:
        type: number
      sku:
        type: string
    type: object
  service.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  service.TokenResponse:
    properties:
      refresh_token:
        type: string
      token:
        type: string
    type: object
  service.UpdateUserRequest:
    properties:
      email:
        type: string
      phone:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  service.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      phone:
        type: string
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is an API for managing users (CRUD) with Clean Architecture.
  title: User Management API
  version: "1.0"
paths:
  /api/orders:
    post:
      consumes:
      - application/json
      description: Creates an EXPORT or IMPORT order manipulating stock via strict
        ACID transactions and broadcasting WS updates
      parameters:
      - description: Create Order Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/service.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create inventory order
      tags:
      - inventory
  /api/products:
    get:
      description: Retrieves list of products with current stock
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.ProductResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get products
      tags:
      - inventory
  /login:
    post:
      consumes:
      - application/json
      description: Authenticates a user by email and password, returning a JWT token
      parameters:
      - description: Login Credentials
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/service.LoginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.TokenResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: Login user
      tags:
      - auth
  /refresh:
    post:
      consumes:
      - application/json
      description: Issues a new access token and refresh token using a valid refresh
        token
      parameters:
      - description: Refresh Token
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/service.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.TokenResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: Refresh token
      tags:
      - auth
  /users:
    get:
      description: Retrieves a paginated list of users
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Number of items per page (default 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Creates a new user validating constraints and hashing password
      parameters:
      - description: Create User Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/service.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create a new user
      tags:
      - users
  /users/{id}:
    delete:
      description: Soft deletes a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
    get:
      description: Fetch a single user's detail by their UUID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.UserResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Updates a user's details excluding password
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Update User Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/service.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"